+++ changelog.md
@@ -7,6 +7,7 @@
     - General cleanup
     - Reunite config classes using Lombok. This breaks part of the api
     - Get rid of Guava and math3 dependency
+    - Incorporate postprocessing directly into recursive clustering step
 
 ## v1.3.1